visualizer/requirements.txt
Flask
kafka-python
retry